Refactor `CertificationInformation` to its own file/package
