+++ wallet-provider-service/src/main/kotlin/domain/certificationinformation/CertificationInformation.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) 2025-2026 European Commission
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package eu.europa.ec.eudi.walletprovider.domain.certificationinformation
+
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.JsonElement
+import kotlinx.serialization.json.JsonObject
+import kotlinx.serialization.json.JsonPrimitive
+
+@JvmInline
+@Serializable
+value class CertificationInformation(
+    val value: JsonElement,
+) {
+    init {
+        require((value is JsonPrimitive && value.isString && value.content.isNotBlank()) || value is JsonObject)
+    }
+
+    override fun toString(): String = value.toString()
+}
